+++ spectrochempy/core/dataset/coordset.py
@@ -15,7 +15,8 @@ import warnings
 import uuid
 
 import numpy as np
-from traitlets import HasTraits, List, Bool, Unicode, observe, All, validate, default, Dict
+from traitlets import HasTraits, List, Bool, Unicode, observe, All, validate, \
+    default, Dict, Int
 
 from spectrochempy.core.dataset.ndarray import NDArray, DEFAULT_DIM_NAME
 from spectrochempy.core.dataset.coord import Coord, LinearCoord
@@ -43,6 +44,9 @@ class CoordSet(HasTraits):
     _copy = Bool(False)
     _sorted = Bool(True)
     _html_output = Bool(False)
+
+    # default coord index
+    _default = Int(0)
 
     # ------------------------------------------------------------------------------------------------------------------
     # initialization
@@ -172,9 +176,8 @@ class CoordSet(HasTraits):
                         # use the provided list of dims
                         coord.name = dims.pop(-1)
 
-                self._append(
-                    coord)  # append the coord (but instead of append,  # use assignation -in _append - to fire the
-                # validation process )
+                self._append(coord)  # append the coord (but instead of append,
+                # use assignation -in _append - to fire the validation process )
 
         # now evaluate keywords argument
         # ------------------------------
@@ -323,6 +326,9 @@ class CoordSet(HasTraits):
     # ..................................................................................................................
     @property
     def coords(self):
+        """
+        list -Coordinates in the coordset
+        """
         return self._coords
 
     # ..................................................................................................................
@@ -344,8 +350,8 @@ class CoordSet(HasTraits):
     # ..................................................................................................................
     @property
     def is_empty(self):
-        """bool - True if there is no coords defined (readonly
-        property).
+        """
+        bool - True if there is no coords defined.
         """
         if self._coords:
             return len(self._coords) == 0
@@ -355,8 +361,8 @@ class CoordSet(HasTraits):
     # ..................................................................................................................
     @property
     def is_same_dim(self):
-        """bool - True if the coords define a single dimension (readonly
-        property).
+        """
+        bool - True if the coords define a single dimension
         """
         return self._is_same_dim
 
@@ -409,6 +415,19 @@ class CoordSet(HasTraits):
     # ------------------------------------------------------------------------------------------------------------------
     # Mutable Properties
     # ------------------------------------------------------------------------------------------------------------------
+
+    @property
+    def default(self):
+        """
+        Coord - default coordinates
+        """
+        return self[self._default]
+
+    @property
+    def data(self):
+        # in case data is called on a coordset for dimension with multiple coordinates
+        # return the first coordinates
+        return self.default.data
 
     # ..................................................................................................................
     @property
@@ -486,6 +505,13 @@ class CoordSet(HasTraits):
         return keys
 
     # ..................................................................................................................
+    def select(self, val):
+        """
+        Select the default coord index
+        """
+        self._default = min(max(0, int(val) - 1), len(self.names))
+
+    #..................................................................................................................
     def set(self, *args, **kwargs):
         """
         Set one or more coordinates in the current CoordSet
@@ -970,6 +996,8 @@ class CoordSet(HasTraits):
     def __deepcopy__(self, memo):
         coords = self.__class__(tuple(cpy.deepcopy(ax, memo=memo) for ax in self), keepnames=True)
         coords.name = self.name
+        coords._is_same_dim = self._is_same_dim
+        coords._default = self._default
         return coords
 
     # ..................................................................................................................
@@ -977,9 +1005,9 @@ class CoordSet(HasTraits):
         coords = self.__class__(tuple(cpy.copy(ax) for ax in self), keepnames=True)
         # name must be changed
         coords.name = self.name
-        # ans is_same_dim too for coordset
-        if self.implements('CoordSet'):
-            coords._is_same_dim = self.is_same_dim
+        # and is_same_dim and default for coordset
+        coords._is_same_dim = self._is_same_dim
+        coords._default = self._default
         return coords
 
         # ..................................................................................................................