+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # What\'s new
 
+## Version 0.2.14
+
+**NEW FEATURES**
+
+* a default coordiate can now be selected for for multiple coordinates dimensions.
+
+**BUGS FIXED**
+
+* alignment along several dimensions (issue #248)
+
+## Version 0.2.13
+
+**BUGS FIXED**
+
+* Solved the problem that reading of experimental datasets was too slow in v.0.2.12.
+
 ## Version 0.2.12
 
 **BUGS FIXED**